requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.12
alembic==1.8.1
apex==0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
cairocffi==1.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
class-resolver==0.3.10
click==8.1.3
cloudpickle==2.1.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.32
data==0.4
databricks-cli==0.17.0
debugpy==1.6.2
decorator==4.4.2
deepchem==2.7.1
defusedxml==0.7.1
dgl-cu110==0.6.1
dgl-cu116==0.8.2.post1
dglgo==0.0.2
dgllife==0.3.1
docker==5.0.3
docopt==0.6.2
docutils==0.18.1
drugood==0.0.1
entrypoints==0.4
executing==0.8.3
faiss-gpu==1.7.2
fastjsonschema==2.16.1
filelock==3.9.0
Flask==2.1.3
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2022.11.0
funcsigs==1.0.2
future==0.18.2
fuzzywuzzy==0.18.0
gensim==4.2.0
geometric2dr==0.5.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.15.0
google-auth-oauthlib==0.4.6
GraKeL==0.1.9
greenlet==1.1.2
grpcio==1.51.1
gunicorn==20.1.0
huggingface-hub==0.11.1
hyperopt==0.2.7
idna==3.3
igraph==0.10.4
imageio==2.22.4
imagesize==1.4.1
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
importlib-resources==5.8.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterthemes==0.20.0
karateclub==1.3.3
kiwisolver==1.4.4
latex==0.7.0
lesscpy==0.15.0
Levenshtein==0.20.8
lightgbm==3.3.3
lightning-utilities==0.4.2
littleutils==0.2.2
llvmlite==0.39.1
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.27.0
mlflow-export-import==1.0.7
mmcv==1.5.0
mpmath==1.2.1
multidict==6.0.3
munkres==1.1.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.12
numba==0.56.4
numpy==1.22.4
numpydoc==1.4.0
oauthlib==3.2.0
ogb==1.3.3
opencv-python==4.7.0.68
outdated==0.2.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkg_resources==0.0.0
ply==3.11
prettytable==3.6.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py3Dmol==2.0.1.post1
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyg-lib==0.2.0+pt112cu116
PyGCL==0.1.2
Pygments==2.12.0
PyGSP==0.5.1
PyJWT==2.4.0
pynauty==2.8.6
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.18.1
PyTDC==0.3.8
python-dateutil==2.8.2
python-Levenshtein==0.20.8
python-louvain==0.16
pytorch-lightning==1.8.4.post0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
querystring-parser==1.2.4
rapidfuzz==2.13.5
rdkit==2022.9.5
rdkit-pypi==2022.3.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.2.2
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0.post4
smart-open==6.2.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.3.0
sympy==1.11.1
tabulate==0.8.10
tempdir==0.7.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.15.0
test-tube==0.7.5
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
torch==1.12.0+cu116
torch-cluster==1.6.1
torch-geometric==2.3.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-spline-conv==1.2.2
torchaudio==0.12.0+cu116
torchmetrics==0.11.0
torchvision==0.13.0+cu116
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.25.1
typer==0.6.1
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wilds==2.0.0
xgboost==1.6.2
yapf==0.32.0
yarg==0.1.9
yarl==1.8.2
zipp==3.8.1